Refactor: 전적 가져오기 -> 5개 가져오기

+++ Search-service/src/main/java/com/ggwp/searchservice/summoner/dto/RequestCreateSummonerDto.java
@@ -23,7 +23,7 @@ public class RequestCreateSummonerDto {
 
     private Long revisionDate;
 
-    private Long summonerLevel;
+    private int summonerLevel;
 
     //    private String accountId;
